refactore: move wallet components out from the bar
